Demo: Add addon tabs, shuffle order
